+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+- Added `defun_prefix` option to `#[module]`.
+- Fixed `#[defun]` not handling raw identifiers correctly.
 
 ## [0.10.1] - 2019-07-18
 - Made `Vector::get` generic by return type.