+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [0.10.3] - 2019-07-24
 - Made `#[defun]` function signatures display correctly in `help-mode` and `helpful-mode`.
 
 ## [0.10.2] - 2019-07-24
@@ -85,7 +87,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [0.2.0] - 2018-01-04
 New reworked version
 
-[Unreleased]: https://github.com/ubolonton/emacs-module-rs/compare/0.10.2...HEAD
+[Unreleased]: https://github.com/ubolonton/emacs-module-rs/compare/0.10.3...HEAD
+[0.10.3]: https://github.com/ubolonton/emacs-module-rs/compare/0.10.2...0.10.3
 [0.10.2]: https://github.com/ubolonton/emacs-module-rs/compare/0.10.1...0.10.2
 [0.10.1]: https://github.com/ubolonton/emacs-module-rs/compare/0.10.0...0.10.1
 [0.10.0]: https://github.com/ubolonton/emacs-module-rs/compare/0.9.0...0.10.0